# app/Http/Controllers/AuthController.php
<?php

namespace App\Http\Controllers;

use App\Http\Requests\LoginRequest;
use App\Models\User;
use Exception;
use Illuminate\Http\Response;

class AuthController extends Controller
{
    private $userModel;

    public function __construct(User $user)
    {
        $this->userModel = $user;
    }

    /**
     * @OA\Post(
     *     path="/api/auth/login",
     *     summary="Autenticação do usuário",
     *     tags={"Autenticação"},
     *     @OA\RequestBody(
     *         required=true,
     *         description="Credenciais para login",
     *         @OA\JsonContent(
     *             required={"email","password"},
     *             @OA\Property(property="email", type="string", format="email", example="usuario@email.com"),
     *             @OA\Property(property="password", type="string", format="password", example="senha123")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Login bem-sucedido",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(
     *                 property="data",
     *                 type="object",
     *                 @OA\Property(property="token", type="string", example="e62jsf283h...")
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Credenciais inválidas",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="false"),
     *             @OA\Property(property="error", type="string", example="Não autorizado!")
     *         )
     *     )
     * )
     */
    public function login(LoginRequest $request): object
    {
        try {
            $input = $request->all();

            $userExist = $this->userModel->getUserByEmail($input['email']);

            if (empty($userExist)) {
                return response()->json(['success' => false, 'error' => 'Usuário não encontrado!'], Response::HTTP_NOT_FOUND);
            }

            $token = auth()->attempt($input);

            if (!$token) {
                return response()->json(['success' => false, 'error' => 'Não autorizado!'], Response::HTTP_UNAUTHORIZED);
            }

            return response()->json(['success' => true, 'data' => ['token' => $token]], Response::HTTP_OK);
        } catch (Exception $e) {
            return response()->json(['succes' => false, 'error' => 'Erro interno, contate o suporte'], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }

    /**
     * @OA\Get(
     *     path="/api/auth/me",
     *     summary="Dados usuário autenticado",
     *     security={{"bearerAuth":{}}},
     *     tags={"Autenticação"},
     *     @OA\Response(
     *         response=200,
     *         description="Dados encontrados!",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(
     *                  property="data",
     *                  type="array",
     *                  @OA\Items(
     *                      type="object",
     *                      @OA\Property(property="id", type="integer", example=1),
     *                      @OA\Property(property="name", type="string", example="admin"),
     *                      @OA\Property(property="email", type="string", example="example@email.com"),
     *                      @OA\Property(property="created_at", type="string", example="2025-08-29 00:00:00"),
     *                      @OA\Property(property="updated_at", type="string", example="2025-08-29 00:00:00")
     *                  )
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Credenciais inválidas",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="false"),
     *             @OA\Property(property="error", type="string", example="Não autorizado!")
     *         )
     *     )
     * )
     */
    public function me(): object
    {
        try {
            $user = auth('api')->user();
            return response()->json(['success' => true, 'data' => $user], Response::HTTP_OK);
        } catch (Exception $e) {
            return response()->json(['succes' => false, 'error' => 'Erro interno, contate o suporte'], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }

    /**
     * @OA\Get(
     *     path="/api/auth/logout",
     *     summary="Logout",
     *     security={{"bearerAuth":{}}},
     *     tags={"Autenticação"},
     *     @OA\Response(
     *         response=200,
     *         description="Deslogado com sucesso!",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(
     *                  property="data",
     *                  type="array",
     *                  @OA\Items(
     *                      type="object",
     *                      @OA\Property(property="message", type="string", example="Sessão finalizada!")
     *                  )
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Credenciais inválidas",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example="false"),
     *             @OA\Property(property="error", type="string", example="Não autorizado!")
     *         )
     *     )
     * )
     */
    public function logout(): object
    {
        try {
            auth()->logout(true);
            return response()->json(['success' => true, 'data' => ['message' => 'Sessão finalizada!']], Response::HTTP_OK);
        } catch (Exception $e) {
            return response()->json(['succes' => false, 'error' => 'Erro interno, contate o suporte'], Response::HTTP_INTERNAL_SERVER_ERROR);
        }
    }
}